feat: evaluate answers

+++ Croco.Mental/Croco.Mental.Domain/Models/RecommendationResponse.cs
@@ -6,8 +6,8 @@ namespace Croco.Mental.Domain.Models
 {
     public sealed class RecommendationResponse
     {
-        public long GoodScore { get; set; }
-        public long BadScore { get; set; }
-        public long Score { get; set; }
+        public int GoodScore { get; set; }
+        public int BadScore { get; set; }
+        public int Score { get; set; }
     }
 }
